% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_corrected_text.R
\name{generate_corrected_text}
\alias{generate_corrected_text}
\title{Generate corrected text}
\usage{
generate_corrected_text(text_original)
}
\arguments{
\item{text_original}{smaller amount of text}
}
\value{
Corrected text data
}
\description{
This function takes in a text input, detects errors and corrects them.
}
