% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_external_consistency_score.R
\name{generate_external_consistency_score}
\alias{generate_external_consistency_score}
\title{Generate external consistency score}
\usage{
generate_external_consistency_score(text_to_check)
}
\arguments{
\item{text_to_check}{smaller amount of text}
}
\value{
A list of two tibbles that contain external consistency and unexpected words
}
\description{
This function takes a small amount of text and generates an
external consistency score
}
\examples{
aRianna::generate_external_consistency_score("there was no possibiliti")

}
