% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_internal_consistency_score.R
\name{generate_internal_consistency_score}
\alias{generate_internal_consistency_score}
\title{Generate internal consistency score}
\usage{
generate_internal_consistency_score(text_to_check, consistency_dataset)
}
\arguments{
\item{text_to_check}{smaller amount of text}

\item{consistency_dataset}{dataset created from make_internal_consistency_dataset}
}
\value{
A list of two tibbles that contain internal consistency and unexpected words
}
\description{
This function takes a small amount of text and generates an
internal consistency score
}
\examples{
body_of_text <- "There was no possibility of taking a walk that day.
We had been wandering, indeed, in the leafless shrubbery an hour in
the morning; but since dinner (Mrs. Reed, when there was no company,
dined early) the cold winter wind had brought with it clouds so sombre,
and a rain so penetrating, that further out-door exercise was now out
of the question."

text_to_evaluate <- "when there was na company"

consistency_dataset <- aRianna::make_internal_consistency_dataset(body_of_text)

aRianna::generate_internal_consistency_score(text_to_evaluate, consistency_dataset)

}
\seealso{
\code{make_internal_consistency_dataset} for consistency_dataset generation
}
