% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_internal_consistency_dataset.R
\name{make_internal_consistency_dataset}
\alias{make_internal_consistency_dataset}
\title{Create an internal consistency dataset}
\usage{
make_internal_consistency_dataset(body_of_text)
}
\arguments{
\item{body_of_text}{larger amount of text}
}
\value{
A collection of n-gram tokens in a tibble
}
\description{
This function takes a large text input creates the dataset that
is needed.
}
\examples{
body_of_text <- "There was no possibility of taking a walk that day.
We had been wandering, indeed, in the leafless shrubbery an hour in
the morning; but since dinner (Mrs. Reed, when there was no company,
dined early) the cold winter wind had brought with it clouds so sombre,
and a rain so penetrating, that further out-door exercise was now out
of the question."
aRianna::make_internal_consistency_dataset(body_of_text)

}
