% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_proportion_func.R
\name{number_proportion}
\alias{number_proportion}
\title{kctest}
\usage{
number_proportion(numbers)
}
\arguments{
\item{group}{of number in vector}
}
\value{
a matrix of the numbers and their proportion values
}
\description{
This number_proportion() function takes a group of numbers and caculates the proportion of each number to the sum of all the numbers
}
\examples{
numbers <- c(2, 5, 3, 9, 8, 11, 6)
number_proportion(numbers)

}
