% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_external_consistency_score.R
\name{generate_external_consistency_score}
\alias{generate_external_consistency_score}
\title{generate_internal_consistency_score}
\usage{
generate_external_consistency_score(text_to_check)
}
\arguments{
\item{text_to_check}{smaller amount of text}
}
\value{

}
\description{
This function takes a small amount of text and generates an
internal consistency score
}
\examples{


}
