% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_internal_consistency_score.R
\name{generate_internal_consistency_score}
\alias{generate_internal_consistency_score}
\title{generate_internal_consistency_score}
\usage{
generate_internal_consistency_score(text_to_check, consistency_dataset)
}
\arguments{
\item{text_to_check}{smaller amount of text}

\item{consistency_dataset}{dataset created from make_internal_consistency_dataset}
}
\value{

}
\description{
This function takes a small amount of text and generates an
internal consistency score
}
\examples{


}
