% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_internal_consistency_dataset.R
\name{make_internal_consistency_dataset}
\alias{make_internal_consistency_dataset}
\title{make_internal_consistency_dataset}
\usage{
make_internal_consistency_dataset(body_of_text)
}
\arguments{
\item{body_of_text}{larger amount of text}
}
\value{

}
\description{
This function takes a large text input creates the dataset that
is needed.
}
\examples{


}
